requirements.txt
torch>=2.2.0
transformers>=4.38.0
peft>=0.10.0
trl>=0.8.0
bitsandbytes>=0.43.0
accelerate>=0.28.0
datasets>=2.18.0
scipy
pyyaml
requests